Add form validation, move copyUI to own file
