+++ package.json
@@ -552,7 +552,7 @@
     "webpack-dev": "webpack --mode development --watch"
   },
   "devDependencies": {
-    "@halcyontech/vscode-ibmi-types": "^1.8.0",
+    "@halcyontech/vscode-ibmi-types": "^2.0.0",
     "@types/glob": "^7.1.3",
     "@types/node": "14.x",
     "@types/vscode": "^1.70.0",
